+++ application/src/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 
 import Plot from "./components/Plot";
-import TabNavigator from "./containers/TabNavigator";
+import TabNavigator from "./components/TabNavigator";
 
 function App() {
   const [plot, setPlot] = useState("hack-813290_960_720.jpg");
